Make the language server binary executable after download (#20)
Resolves #19
Draft because I'm unsure this repo is the correct place to fix the
issue. Maybe it should be `zed::download_file`'s job to carry over file
permissions when extracting a zip file.

Note that `zed::make_file_executable` is a noop on non-unix platforms,
so it shouldn't be an issue to call it here.