Fix binary path caching
The extension was storing the path to **only** the first language server it downloaded.
Still probably shouldn't be running with more than one, but this at least fixes
the argument error that arises from a freshly installed, unconfigured setup.

In brief, the error was arising due to `kotlin-lsp` args getting sent to
`kotlin-language-server`. `kotlin-language-server` doesn't support `--stdio`.