Fix binary path caching (#37)
Resolves #31
Resolves #33

The extension was storing the path to **only** the first language server
it downloaded. I can't imagine you'd want to be running with more than
one, but this at least fixes the argument error that arises from a
freshly installed, unconfigured setup.

In brief, the error was arising due to `kotlin-lsp` args getting sent to
`kotlin-language-server`. `kotlin-language-server` doesn't support the
`--stdio` argument. It also meant that should it have worked there
would've been two instances of `kotlin-language-server` running.